pyproject.toml
[tool.poetry]
name = "rfid_lab"
version = "0.1.0"
description = ""
authors = [

]
maintainers = [

]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.100.0"
uvicorn = { version = "^0.22.0", extras = ["standard"] }
gunicorn = "^21.2.0"
pydantic = "^2"
pydantic-settings = "^2"
yarl = "^1.9.2"
ujson = "^5.8.0"
tortoise-orm = "^0.19.3"
aerich = "^0.7.1"
aiosqlite = "<0.19.0"
httptools = "^0.6.0"
loguru = "^0.7.0"
pyserial = "^3.5"


[tool.poetry.dev-dependencies]
pytest = "^7.2.1"
flake8 = "~4.0.1"
mypy = "^1.1.1"
isort = "^5.11.4"
pre-commit = "^3.0.1"
wemake-python-styleguide = "^0.17.0"
black = "^22.12.0"
autoflake = "^1.6.1"
pytest-cov = "^4.0.0"
anyio = "^3.6.2"
pytest-env = "^0.8.1"
asynctest = "^0.13.0"
nest-asyncio = "^1.5.6"
httpx = "^0.23.3"

[tool.isort]
profile = "black"
multi_line_output = 3
src_paths = ["rfid_lab",]

[tool.mypy]
strict = true
ignore_missing_imports = true
allow_subclassing_any = true
allow_untyped_calls = true
pretty = true
show_error_codes = true
implicit_reexport = true
allow_untyped_decorators = true
warn_unused_ignores = false
warn_return_any = false
namespace_packages = true

[tool.pytest.ini_options]
filterwarnings = [
    "error",
    "ignore::DeprecationWarning",
    "ignore:.*unclosed.*:ResourceWarning",
]
env = [
    "RFID_LAB_ENVIRONMENT=pytest",
    "RFID_LAB_DB_FILE=test_db.sqlite3",
]
[tool.aerich]
tortoise_orm = "rfid_lab.db.config.TORTOISE_CONFIG"
location = "./rfid_lab/db/migrations"
src_folder = "./rfid_lab"

[fastapi-template.options]
project_name = "rfid_lab"
api_type = "rest"
enable_redis = "None"
enable_rmq = "None"
ci_type = "github"
enable_migrations = "True"
enable_taskiq = "None"
enable_kube = "None"
kube_name = "rfid-lab"
enable_routers = "True"
enable_kafka = "None"
enable_loguru = "True"
traefik_labels = "None"
add_dummy = "True"
orm = "tortoise"
self_hosted_swagger = "None"
prometheus_enabled = "None"
sentry_enabled = "None"
otlp_enabled = "None"
pydanticv1 = "None"
gunicorn = "True"
add_users = "None"
cookie_auth = "None"
jwt_auth = "None"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
